refactor: move pingUser api to notifications, getStream types to the service
