--- src/api/inAppNotifications/index.ts
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: ice License 1.0
-
-import {pingUser} from './pingUser';
-
-export const inAppNotifications = Object.freeze({
-  pingUser,
-});
