+++ src/api/index.ts
@@ -2,7 +2,6 @@
 
 import {achievements} from '@api/achievements';
 import {devices} from '@api/devices';
-import {inAppNotifications} from '@api/inAppNotifications';
 import {news} from '@api/news';
 import {notifications} from '@api/notifications';
 import {referrals} from '@api/referrals';
@@ -23,7 +22,6 @@ export const Api = Object.freeze({
   devices,
   tokenomics,
   news,
-  inAppNotifications,
   statusNotice,
   tasks,
   time,
