+++ src/store/modules/InAppNotifications/reducer/index.ts
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: ice License 1.0
 
-import {Activity} from '@api/inAppNotifications/types';
+import {Activity} from '@services/getStream/types';
 import {InAppNotificationActions} from '@store/modules/InAppNotifications/actions';
 import produce from 'immer';
