+++ src/store/modules/InAppNotifications/sagas/loadInAppNotificationsSaga.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: ice License 1.0
 
-import {Activity} from '@api/inAppNotifications/types';
 import {loadAnnouncements, loadInAppNotifications} from '@services/getstream';
+import {Activity} from '@services/getStream/types';
 import {InAppNotificationActions} from '@store/modules/InAppNotifications/actions';
 import {getErrorMessage} from '@utils/errors';
 import {call, put, SagaReturnType} from 'redux-saga/effects';
